tst_QDockWidget: use local context for timer
If we use 'this' as context it might still try to invoke
the timer after we have left the function, leading to
stack-use-after-return.

To avoid doing an in-depth dive if it's okay to use mainWindow as the
context or not, I just added a new local QObject.

Amends 9ff40b59da58160dc26c54204a615a2456e07405

Pick-to: 6.5
Change-Id: I2c3bdc1eb06731d9c38979610303876c2748fb73
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 26ded9dedf3f35e93b81bd0281fdd0bd30acb9d4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>